Support Cylinder shape, merge redundant class
