+++ BuildOutputs/Database_Summary.txt
@@ -1,11 +1,11 @@
 MD5 Checksum
-3D2E402E586C18CD8ACFC89EE490BC5D
+F2BD0562AE6685FA714BD66D32E89576
 
 Size
 456152 KB
 
 Built At UTC
-31/07/2025 23:00:02
+11/08/2025 21:20:30
 
 CSV Imported with SQLite Version
 3.50.2 2025-06-28 14:00:48 2af157d77fb1304a74176eaee7fbc7c7e932d946bf25325e9c26c91db19e3079 (64-bit)
@@ -252,7 +252,8 @@ Unknown,1
 
 Primitive Shapes
 ,4253707
-Box,124183
+Box,124049
+Cylinder,134
 Ellipsoid,4
 Line,1965
 Plane,359
